feat: registration comoponent structure
