+++ app/register/page.tsx
@@ -1,0 +1,11 @@
+import Register from "@/ui-components/authentication/register/register";
+
+const RegisterPage = () => {
+  return (
+    <>
+      <Register />
+    </>
+  );
+};
+
+export default RegisterPage;
