+++ ui-components/authentication/login/login-providers.tsx
@@ -1,0 +1,12 @@
+import Login from "./login";
+
+
+const LoginOptions = () => {
+  return (
+    <div className="flex w-full h-full  justify-center">
+      <Login />
+    </div>
+  );
+};
+
+export default LoginOptions;
